Surface culling and owning
Each object can mark outgoing rays as "owned" by it, so that only this object's
surfaces will be checked for intersection with these rays at the next round.

Furthermore, for each ray going out of an object, the object can say which of
its own surfaces can be ignored at the next iteration. A default implementation
and a specialized one for one_sided_mirror are provided.